Moved query builder test harness into it's own folder
